+++ google/cloud/spanner_dbapi/connection.py
@@ -20,6 +20,7 @@ import warnings
 from google.api_core.exceptions import Aborted
 from google.api_core.gapic_v1.client_info import ClientInfo
 from google.cloud import spanner_v1 as spanner
+from google.cloud.spanner_v1 import RequestOptions
 from google.cloud.spanner_v1.session import _get_retry_delay
 from google.cloud.spanner_v1.snapshot import Snapshot
 
@@ -103,6 +104,7 @@ class Connection:
         self._own_pool = True
         self._read_only = read_only
         self._staleness = None
+        self.request_priority = None
 
     @property
     def autocommit(self):
@@ -180,6 +182,21 @@ class Connection:
                 "Commit or rollback the current transaction and try again."
             )
         self._read_only = value
+
+    @property
+    def request_options(self):
+        """Options for the next SQL operations.
+
+        Returns:
+            google.cloud.spanner_v1.RequestOptions:
+                Request options.
+        """
+        if self.request_priority is None:
+            return
+
+        req_opts = RequestOptions(priority=self.request_priority)
+        self.request_priority = None
+        return req_opts
 
     @property
     def staleness(self):
@@ -435,7 +452,7 @@ class Connection:
 
         if statement.is_insert:
             _execute_insert_heterogenous(
-                transaction, ((statement.sql, statement.params),)
+                transaction, ((statement.sql, statement.params),), self.request_options
             )
             return (
                 iter(()),
@@ -447,6 +464,7 @@ class Connection:
                 statement.sql,
                 statement.params,
                 param_types=statement.param_types,
+                request_options=self.request_options,
             ),
             ResultsChecksum() if retried else statement.checksum,
         )