+++ samples/samples/snippets_test.py
@@ -17,6 +17,7 @@ import uuid
 
 from google.api_core import exceptions
 from google.cloud import spanner
+from google.cloud.spanner_admin_database_v1.types.common import DatabaseDialect
 import pytest
 from test_utils.retry import RetryErrors
 
@@ -46,6 +47,16 @@ retry_429 = RetryErrors(exceptions.ResourceExhausted, delay=15)
 @pytest.fixture(scope="module")
 def sample_name():
     return "snippets"
+
+
+@pytest.fixture(scope="module")
+def database_dialect():
+    """Spanner dialect to be used for this sample.
+
+    The dialect is used to initialize the dialect for the database.
+    It can either be GoogleStandardSql or PostgreSql.
+    """
+    return DatabaseDialect.GOOGLE_STANDARD_SQL
 
 
 @pytest.fixture(scope="module")
@@ -95,6 +106,21 @@ def default_leader():
     return "us-east4"
 
 
+@pytest.fixture(scope="module")
+def user_managed_instance_config_name(spanner_client):
+    name = f"custom-python-samples-config-{uuid.uuid4().hex[:10]}"
+    yield name
+    snippets.delete_instance_config(
+      "{}/instanceConfigs/{}".format(spanner_client.project_name, name)
+    )
+    return
+
+
+@pytest.fixture(scope="module")
+def base_instance_config_id(spanner_client):
+    return "{}/instanceConfigs/{}".format(spanner_client.project_name, "nam7")
+
+
 def test_create_instance_explicit(spanner_client, create_instance_id):
     # Rather than re-use 'sample_isntance', we create a new instance, to
     # ensure that the 'create_instance' snippet is tested.
@@ -114,7 +140,8 @@ def test_create_database_explicit(sample_instance, create_database_id):
 def test_create_instance_with_processing_units(capsys, lci_instance_id):
     processing_units = 500
     retry_429(snippets.create_instance_with_processing_units)(
-        lci_instance_id, processing_units,
+      lci_instance_id,
+      processing_units,
     )
     out, _ = capsys.readouterr()
     assert lci_instance_id in out
@@ -125,10 +152,10 @@ def test_create_instance_with_processing_units(capsys, lci_instance_id):
 
 
 def test_create_database_with_encryption_config(
-    capsys, instance_id, cmek_database_id, kms_key_name
+  capsys, instance_id, cmek_database_id, kms_key_name
 ):
     snippets.create_database_with_encryption_key(
-        instance_id, cmek_database_id, kms_key_name
+      instance_id, cmek_database_id, kms_key_name
     )
     out, _ = capsys.readouterr()
     assert cmek_database_id in out
@@ -148,6 +175,42 @@ def test_list_instance_config(capsys):
     assert "regional-us-central1" in out
 
 
+@pytest.mark.dependency(name="create_instance_config")
+def test_create_instance_config(
+  capsys, user_managed_instance_config_name, base_instance_config_id
+):
+    snippets.create_instance_config(
+      user_managed_instance_config_name, base_instance_config_id
+    )
+    out, _ = capsys.readouterr()
+    assert "Created instance configuration" in out
+
+
+@pytest.mark.dependency(depends=["create_instance_config"])
+def test_update_instance_config(capsys, user_managed_instance_config_name):
+    snippets.update_instance_config(user_managed_instance_config_name)
+    out, _ = capsys.readouterr()
+    assert "Updated instance configuration" in out
+
+
+@pytest.mark.dependency(depends=["create_instance_config"])
+def test_delete_instance_config(capsys, user_managed_instance_config_name):
+    spanner_client = spanner.Client()
+    snippets.delete_instance_config(
+      "{}/instanceConfigs/{}".format(
+        spanner_client.project_name, user_managed_instance_config_name
+      )
+    )
+    out, _ = capsys.readouterr()
+    assert "successfully deleted" in out
+
+
+def test_list_instance_config_operations(capsys):
+    snippets.list_instance_config_operations()
+    out, _ = capsys.readouterr()
+    assert "List instance config operations" in out
+
+
 def test_list_databases(capsys, instance_id):
     snippets.list_databases(instance_id)
     out, _ = capsys.readouterr()
@@ -155,15 +218,15 @@ def test_list_databases(capsys, instance_id):
 
 
 def test_create_database_with_default_leader(
-    capsys,
-    multi_region_instance,
-    multi_region_instance_id,
-    default_leader_database_id,
-    default_leader,
+  capsys,
+  multi_region_instance,
+  multi_region_instance_id,
+  default_leader_database_id,
+  default_leader,
 ):
     retry_429 = RetryErrors(exceptions.ResourceExhausted, delay=15)
     retry_429(snippets.create_database_with_default_leader)(
-        multi_region_instance_id, default_leader_database_id, default_leader
+      multi_region_instance_id, default_leader_database_id, default_leader
     )
     out, _ = capsys.readouterr()
     assert default_leader_database_id in out
@@ -171,15 +234,15 @@ def test_create_database_with_default_leader(
 
 
 def test_update_database_with_default_leader(
-    capsys,
-    multi_region_instance,
-    multi_region_instance_id,
-    default_leader_database_id,
-    default_leader,
+  capsys,
+  multi_region_instance,
+  multi_region_instance_id,
+  default_leader_database_id,
+  default_leader,
 ):
     retry_429 = RetryErrors(exceptions.ResourceExhausted, delay=15)
     retry_429(snippets.update_database_with_default_leader)(
-        multi_region_instance_id, default_leader_database_id, default_leader
+      multi_region_instance_id, default_leader_database_id, default_leader
     )
     out, _ = capsys.readouterr()
     assert default_leader_database_id in out
@@ -193,14 +256,14 @@ def test_get_database_ddl(capsys, instance_id, sample_database):
 
 
 def test_query_information_schema_database_options(
-    capsys,
-    multi_region_instance,
-    multi_region_instance_id,
-    default_leader_database_id,
-    default_leader,
+  capsys,
+  multi_region_instance,
+  multi_region_instance_id,
+  default_leader_database_id,
+  default_leader,
 ):
     snippets.query_information_schema_database_options(
-        multi_region_instance_id, default_leader_database_id
+      multi_region_instance_id, default_leader_database_id
     )
     out, _ = capsys.readouterr()
     assert default_leader in out
@@ -272,7 +335,8 @@ def test_read_write_transaction(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_column"])
 def test_query_data_with_new_column(capsys, instance_id, sample_database):
-    snippets.query_data_with_new_column(instance_id, sample_database.database_id)
+    snippets.query_data_with_new_column(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "SingerId: 1, AlbumId: 1, MarketingBudget: 300000" in out
     assert "SingerId: 2, AlbumId: 2, MarketingBudget: 300000" in out
@@ -312,7 +376,8 @@ def test_add_storing_index(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_storing_index"])
 def test_read_data_with_storing_index(capsys, instance_id, sample_database):
-    snippets.read_data_with_storing_index(instance_id, sample_database.database_id)
+    snippets.read_data_with_storing_index(instance_id,
+                                          sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "300000" in out
 
@@ -334,7 +399,8 @@ def test_add_timestamp_column(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_timestamp_column"])
 def test_update_data_with_timestamp(capsys, instance_id, sample_database):
-    snippets.update_data_with_timestamp(instance_id, sample_database.database_id)
+    snippets.update_data_with_timestamp(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "Updated data" in out
 
@@ -349,14 +415,16 @@ def test_query_data_with_timestamp(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(name="create_table_with_timestamp")
 def test_create_table_with_timestamp(capsys, instance_id, sample_database):
-    snippets.create_table_with_timestamp(instance_id, sample_database.database_id)
+    snippets.create_table_with_timestamp(instance_id,
+                                         sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "Created Performances table on database" in out
 
 
-@pytest.mark.dependency(depends=["create_table_with_datatypes"])
+@pytest.mark.dependency(depends=["create_table_with_timestamp"])
 def test_insert_data_with_timestamp(capsys, instance_id, sample_database):
-    snippets.insert_data_with_timestamp(instance_id, sample_database.database_id)
+    snippets.insert_data_with_timestamp(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "Inserted data." in out
 
@@ -377,7 +445,8 @@ def test_query_with_struct(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["write_struct_data"])
 def test_query_with_array_of_struct(capsys, instance_id, sample_database):
-    snippets.query_with_array_of_struct(instance_id, sample_database.database_id)
+    snippets.query_with_array_of_struct(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "SingerId: 8" in out
     assert "SingerId: 7" in out
@@ -430,14 +499,16 @@ def test_delete_data_with_dml(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_timestamp_column"])
 def test_update_data_with_dml_timestamp(capsys, instance_id, sample_database):
-    snippets.update_data_with_dml_timestamp(instance_id, sample_database.database_id)
+    snippets.update_data_with_dml_timestamp(instance_id,
+                                            sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "2 record(s) updated." in out
 
 
 @pytest.mark.dependency(name="dml_write_read_transaction")
 def test_dml_write_read_transaction(capsys, instance_id, sample_database):
-    snippets.dml_write_read_transaction(instance_id, sample_database.database_id)
+    snippets.dml_write_read_transaction(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "1 record(s) inserted." in out
     assert "FirstName: Timothy, LastName: Campbell" in out
@@ -445,7 +516,8 @@ def test_dml_write_read_transaction(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["dml_write_read_transaction"])
 def test_update_data_with_dml_struct(capsys, instance_id, sample_database):
-    snippets.update_data_with_dml_struct(instance_id, sample_database.database_id)
+    snippets.update_data_with_dml_struct(instance_id,
+                                         sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "1 record(s) updated" in out
 
@@ -466,21 +538,24 @@ def test_query_data_with_parameter(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_column"])
 def test_write_with_dml_transaction(capsys, instance_id, sample_database):
-    snippets.write_with_dml_transaction(instance_id, sample_database.database_id)
+    snippets.write_with_dml_transaction(instance_id,
+                                        sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "Transferred 200000 from Album2's budget to Album1's" in out
 
 
 @pytest.mark.dependency(depends=["add_column"])
 def update_data_with_partitioned_dml(capsys, instance_id, sample_database):
-    snippets.update_data_with_partitioned_dml(instance_id, sample_database.database_id)
+    snippets.update_data_with_partitioned_dml(instance_id,
+                                              sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "3 record(s) updated" in out
 
 
 @pytest.mark.dependency(depends=["insert_with_dml"])
 def test_delete_data_with_partitioned_dml(capsys, instance_id, sample_database):
-    snippets.delete_data_with_partitioned_dml(instance_id, sample_database.database_id)
+    snippets.delete_data_with_partitioned_dml(instance_id,
+                                              sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "6 record(s) deleted" in out
 
@@ -494,13 +569,15 @@ def test_update_with_batch_dml(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(name="create_table_with_datatypes")
 def test_create_table_with_datatypes(capsys, instance_id, sample_database):
-    snippets.create_table_with_datatypes(instance_id, sample_database.database_id)
+    snippets.create_table_with_datatypes(instance_id,
+                                         sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "Created Venues table on database" in out
 
 
 @pytest.mark.dependency(
-    name="insert_datatypes_data", depends=["create_table_with_datatypes"],
+  name="insert_datatypes_data",
+  depends=["create_table_with_datatypes"],
 )
 def test_insert_datatypes_data(capsys, instance_id, sample_database):
     snippets.insert_datatypes_data(instance_id, sample_database.database_id)
@@ -562,7 +639,8 @@ def test_query_data_with_string(capsys, instance_id, sample_database):
 
 
 @pytest.mark.dependency(
-    name="add_numeric_column", depends=["create_table_with_datatypes"],
+  name="add_numeric_column",
+  depends=["create_table_with_datatypes"],
 )
 def test_add_numeric_column(capsys, instance_id, sample_database):
     snippets.add_numeric_column(instance_id, sample_database.database_id)
@@ -578,14 +656,17 @@ def test_update_data_with_numeric(capsys, instance_id, sample_database):
 
 
 @pytest.mark.dependency(depends=["add_numeric_column"])
-def test_query_data_with_numeric_parameter(capsys, instance_id, sample_database):
-    snippets.query_data_with_numeric_parameter(instance_id, sample_database.database_id)
+def test_query_data_with_numeric_parameter(capsys, instance_id,
+                                           sample_database):
+    snippets.query_data_with_numeric_parameter(instance_id,
+                                               sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "VenueId: 4, Revenue: 35000" in out
 
 
 @pytest.mark.dependency(
-    name="add_json_column", depends=["create_table_with_datatypes"],
+  name="add_json_column",
+  depends=["create_table_with_datatypes"],
 )
 def test_add_json_column(capsys, instance_id, sample_database):
     snippets.add_json_column(instance_id, sample_database.database_id)
@@ -602,15 +683,17 @@ def test_update_data_with_json(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["add_json_column"])
 def test_query_data_with_json_parameter(capsys, instance_id, sample_database):
-    snippets.query_data_with_json_parameter(instance_id, sample_database.database_id)
+    snippets.query_data_with_json_parameter(instance_id,
+                                            sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "VenueId: 19, VenueDetails: {'open': True, 'rating': 9}" in out
 
 
 @pytest.mark.dependency(depends=["insert_datatypes_data"])
-def test_query_data_with_timestamp_parameter(capsys, instance_id, sample_database):
+def test_query_data_with_timestamp_parameter(capsys, instance_id,
+                                             sample_database):
     snippets.query_data_with_timestamp_parameter(
-        instance_id, sample_database.database_id
+      instance_id, sample_database.database_id
     )
     out, _ = capsys.readouterr()
     assert "VenueId: 4, VenueName: Venue 4, LastUpdateTime:" in out
@@ -620,7 +703,8 @@ def test_query_data_with_timestamp_parameter(capsys, instance_id, sample_databas
 
 @pytest.mark.dependency(depends=["insert_datatypes_data"])
 def test_query_data_with_query_options(capsys, instance_id, sample_database):
-    snippets.query_data_with_query_options(instance_id, sample_database.database_id)
+    snippets.query_data_with_query_options(instance_id,
+                                           sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "VenueId: 4, VenueName: Venue 4, LastUpdateTime:" in out
     assert "VenueId: 19, VenueName: Venue 19, LastUpdateTime:" in out
@@ -629,7 +713,8 @@ def test_query_data_with_query_options(capsys, instance_id, sample_database):
 
 @pytest.mark.dependency(depends=["insert_datatypes_data"])
 def test_create_client_with_query_options(capsys, instance_id, sample_database):
-    snippets.create_client_with_query_options(instance_id, sample_database.database_id)
+    snippets.create_client_with_query_options(instance_id,
+                                              sample_database.database_id)
     out, _ = capsys.readouterr()
     assert "VenueId: 4, VenueName: Venue 4, LastUpdateTime:" in out
     assert "VenueId: 19, VenueName: Venue 19, LastUpdateTime:" in out